+++ examples/nex_base_demo/config/config.exs
@@ -1,5 +1,0 @@
-import Config
-
-config :nex_base_demo, NexBaseDemo.Repo,
-  adapter: Ecto.Adapters.Postgres,
-  pool_size: 10